Refactor AutoEP presets into registry package
