+++ deepspeed/module_inject/auto_ep_presets/__init__.py
@@ -1,0 +1,3 @@
+# SPDX-License-Identifier: Apache-2.0
+# DeepSpeed Team
+"""AutoEP built-in preset registry package."""
