+++ deepspeed/module_inject/auto_ep_presets/base.py
@@ -1,0 +1,253 @@
+# SPDX-License-Identifier: Apache-2.0
+# DeepSpeed Team
+"""Shared AutoEP preset dataclasses and adapter interface."""
+
+from __future__ import annotations
+
+from dataclasses import dataclass, field
+from typing import Any, Callable, Literal
+
+import torch.nn as nn
+from packaging.version import InvalidVersion, Version
+
+# Sentinel for "not specified in config, use preset default".
+# Unlike None (which means "fused gate+up, no separate w3"), _UNSET means
+# the user did not set the field at all. Compare with `is _UNSET`.
+_UNSET = object()
+
+
+@dataclass
+class MoEModelPreset:
+    """Preset configuration for a known MoE model family."""
+
+    moe_layer_pattern: str
+    router_pattern: str
+    experts_pattern: str
+    expert_storage: Literal["fused_3d", "module_list"]
+    expert_w1: str
+    expert_w2: str
+    expert_w3: str | None
+    num_experts_attr: str
+    top_k_attr: str
+    score_func: Literal["softmax", "sigmoid"]
+    score_apply: Literal["pre", "post"]
+    route_norm: bool
+    gate_bias: bool
+    has_shared_experts: bool = False
+    shared_experts_pattern: str = ""
+    shared_experts_gate_pattern: str = ""
+    autoep_config_defaults: dict[str, Any] = field(default_factory=dict)
+    supports_expert_bias: bool = True
+    unsupported_router_bias_names: tuple[str, ...] = ()
+    preset_adapter: str = "default"
+    hf_model_types: tuple[str, ...] = ()
+    unsupported_hf_model_type_notes: dict[str, str] = field(default_factory=dict)
+    min_transformers_version: str | None = None
+    validated_transformers_versions: str = ""
+    docs_support_notes: str = ""
+
+
+@dataclass
+class MoELayerSpec:
+    """Detected MoE layer specification for a single module in the model."""
+
+    moe_module_name: str
+    model_family: str
+    router_name: str
+    experts_name: str
+    expert_storage: Literal["fused_3d", "module_list"]
+    expert_w1_name: str
+    expert_w2_name: str
+    expert_w3_name: str | None
+    num_experts: int
+    top_k: int
+    hidden_size: int
+    ffn_hidden_size: int
+    score_func: Literal["softmax", "sigmoid"]
+    score_apply: Literal["pre", "post"]
+    route_norm: bool
+    gate_bias: bool
+    return_router_logits: bool
+    router_logits_capture_target: Literal["moe_block", "router", "none"]
+    router_logits_capture_index: int | None
+    router_logits_capture_layer_name: str | None
+    has_shared_experts: bool
+    shared_experts_name: str
+    shared_experts_gate_name: str = ""
+    route_scale: float = 1.0
+    num_expert_groups: int | None = None
+    num_limited_groups: int | None = None
+    group_score_func: Literal["max", "top2_sum"] = "top2_sum"
+    supports_expert_bias: bool = True
+    unsupported_router_bias_names: tuple[str, ...] = ()
+    preset_adapter: str = "default"
+    router_logits_capture_mode: Literal["raw", "post_score"] = "post_score"
+    moe_output_shape: Literal["batched", "flat"] = "batched"
+
+
+@dataclass
+class AutoEPConfig:
+    """User-facing configuration parsed from DS config JSON."""
+
+    enabled: bool = False
+    autoep_size: int = 1
+    preset_model: str | None = None
+    moe_layer_pattern: str | None = None
+    expert_pattern: str | None = None
+    router_pattern: str | None = None
+    use_grouped_mm: bool = True
+    route_norm: bool | None = None
+    route_scale: float = 1.0
+    score_apply: Literal["auto", "pre", "post"] = "auto"
+    combine_impl: Literal["auto", "weighted_sum", "legacy_bmm"] = "auto"
+    num_expert_groups: int | None = None
+    num_limited_groups: int | None = None
+    score_func: Literal["auto", "softmax", "sigmoid"] = "auto"
+    top_k: int | str = "auto"
+    load_balance_coeff: float | None | object = _UNSET
+    routed_scaling_factor: float | str = "auto"
+    expert_w1: str | None = None
+    expert_w2: str | None = None
+    expert_w3: object = _UNSET
+    num_experts_attr: str | None = None
+    top_k_attr: str | None = None
+    has_shared_experts: bool | None = None
+    shared_experts_pattern: str | None = None
+    shared_experts_gate_pattern: str | None = None
+    _load_balance_coeff_explicit: bool = field(default=False, init=False, repr=False)
+
+    def __post_init__(self) -> None:
+        if self.load_balance_coeff is _UNSET:
+            self.load_balance_coeff = 1e-3
+            self._load_balance_coeff_explicit = False
+        else:
+            self._load_balance_coeff_explicit = True
+
+
+@dataclass(frozen=True)
+class GroupRoutingConfig:
+    num_expert_groups: int | None
+    num_limited_groups: int | None
+    group_score_func: Literal["max", "top2_sum"] = "top2_sum"
+
+
+@dataclass(frozen=True)
+class ForwardContract:
+    return_router_logits: bool = False
+    capture_target: Literal["moe_block", "router", "none"] = "none"
+    capture_index: int | None = None
+    capture_layer_name: str | None = None
+    router_logits_capture_mode: Literal["raw", "post_score"] = "post_score"
+    moe_output_shape: Literal["batched", "flat"] = "batched"
+
+
+class AutoEPPresetAdapter:
+    """Default behavior shared by presets without model-specific parser rules."""
+
+    def validate_compatibility(
+        self,
+        preset_name: str,
+        preset: MoEModelPreset,
+        model_config,
+    ) -> None:
+        """Validate public HF compatibility metadata for a selected preset."""
+        model_type = getattr(model_config, "model_type", None) if model_config is not None else None
+        self._validate_hf_model_type(preset_name, preset, model_type)
+        self._validate_transformers_version(preset_name, preset, model_type)
+
+    def _validate_hf_model_type(
+        self,
+        preset_name: str,
+        preset: MoEModelPreset,
+        model_type: str | None,
+    ) -> None:
+        if model_type is None:
+            return
+
+        unsupported_note = preset.unsupported_hf_model_type_notes.get(model_type)
+        if unsupported_note is None:
+            return
+
+        supported = ", ".join(repr(value) for value in preset.hf_model_types) or "none"
+        raise ValueError(f"AutoEP preset '{preset_name}' does not support model_type='{model_type}'. "
+                         f"{unsupported_note} Supported HF model_type value(s): {supported}.")
+
+    def _validate_transformers_version(
+        self,
+        preset_name: str,
+        preset: MoEModelPreset,
+        model_type: str | None,
+    ) -> None:
+        min_version = preset.min_transformers_version
+        if min_version is None or model_type is None:
+            return
+        if not self._requires_transformers_version_validation():
+            return
+        if model_type not in preset.hf_model_types and model_type not in preset.unsupported_hf_model_type_notes:
+            return
+
+        try:
+            installed_version = self._installed_transformers_version()
+        except Exception as exc:
+            raise ValueError(f"AutoEP preset '{preset_name}' for model_type='{model_type}' requires "
+                             f"Transformers >= {min_version}, but transformers could not be imported: {exc}.") from exc
+
+        try:
+            installed = Version(installed_version)
+            minimum = Version(min_version)
+        except InvalidVersion as exc:
+            raise ValueError(f"AutoEP preset '{preset_name}' for model_type='{model_type}' requires "
+                             f"Transformers >= {min_version}, but the installed Transformers version "
+                             f"'{installed_version}' could not be parsed.") from exc
+
+        if installed < minimum:
+            raise ValueError(f"AutoEP preset '{preset_name}' for model_type='{model_type}' requires "
+                             f"Transformers >= {min_version}, but installed transformers=={installed_version}. "
+                             "Upgrade Transformers or choose a preset/model combination supported by the "
+                             "installed Transformers version.")
+
+    def _installed_transformers_version(self) -> str:
+        import transformers
+        return getattr(transformers, "__version__", "unknown")
+
+    def _requires_transformers_version_validation(self) -> bool:
+        # The default adapter also covers non-HF/mock/custom-compatible configs;
+        # specialized HF-only adapters opt in to minimum Transformers checks.
+        return False
+
+    def resolve_route_norm(
+        self,
+        config: AutoEPConfig,
+        preset: MoEModelPreset,
+        model_config,
+    ) -> bool:
+        if config.route_norm is not None:
+            return config.route_norm
+
+        cfg_norm = getattr(model_config, 'norm_topk_prob', None)
+        if cfg_norm is not None:
+            return bool(cfg_norm)
+        return preset.route_norm
+
+    def resolve_group_routing(
+        self,
+        config: AutoEPConfig,
+        model_config,
+    ) -> GroupRoutingConfig:
+        return GroupRoutingConfig(
+            num_expert_groups=config.num_expert_groups,
+            num_limited_groups=config.num_limited_groups,
+        )
+
+    def adjust_forward_contract(self, contract: ForwardContract) -> ForwardContract:
+        return contract
+
+    def retarget_transformers_output_recorders(
+        self,
+        model: nn.Module,
+        spec: MoELayerSpec,
+        replacement: nn.Module,
+        retargeted_keys: set[str],
+        remove_output_capture_hooks: Callable[[nn.Module], int],
+    ) -> None:
+        return
